requirement.txt
asgiref==3.7.2
Django==4.2.6
django-appconf==1.0.5
django-avatar==8.0.0
django-ckeditor==6.7.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
dnspython==2.4.2
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.3
